brightwaybot-main/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv==1.0.0
pydantic==2.5.0
asyncio-mqtt==0.16.1
authlib==1.2.1
tensorflow>=2.16.0
scikit-learn>=1.4.0
joblib==1.3.2
psutil==5.9.6